Proportion and Total Diverted

Monthly propotion and total diverted values for the duration of screw trap observations.

rbdd_prop_div

feat_prop_div

tuol_prop_div

amer_prop_div

moke_prop_div

stan_prop_div

deer_prop_div

rbdd_tot_div

feat_tot_div

tuol_tot_div

amer_tot_div

moke_tot_div

stan_tot_div

deer_tot_div

Format

An object of class tbl_df (inherits from tbl, data.frame) with 216 rows and 3 columns.

Details

Available screw trap data start and end years:

Location 	 start-end

RBDD 	 1999 - 2016

Feather River 	 1998 - 2016

Tuolumne River 	 2007 - 2017

American River 	 2013 - 2016

Battle Creek 	 1998 - 2016

Clear Creek 	 1998 - 2016

Deer Creek 	 1992 - 2010

Mokelumne River 	 1999 - 2015

Proportion Diverted

Values pre September 2003 are CALSIM model output and values after are the median CALSIM value grouped by water year type and month.

For more details on CALSIM see:

use this link within R proportion_diverted

use this link if in a web browser

NOTE: RBDD is represented using the Upper Sacramento River DSM model node. There are no modeled diversions for Battle and Clear Creek.

Total Diverted

Values were estimated by multiplying the measured flow data and the modeled proportion diverted values.