Gaged Flow, Flow Coefficient of Variation, and Temperature

Measured daily and monthly mean flow (cfs) and temperature (°C) from USGS and CDEC gages for the duration of screw trap observations. The coefficient of variation (CV=sd/mean) for monthly flow was also calculated.

rbdd_flow

feather_flow

tuol_flow

amer_flow

moke_flow

stan_flow

battle_flow

clear_flow

deer_flow

rbdd_temp

feather_temp

tuol_temp

amer_temp

moke_temp

stan_temp

battle_temp

clear_temp

deer_temp

rbdd_cv

feather_cv

tuol_cv

amer_cv

moke_cv

stan_cv

battle_cv

clear_cv

deer_cv

rbdd_daily_flow

feather_daily_flow

tuol_daily_flow

amer_daily_flow

moke_daily_flow

stan_daily_flow

battle_daily_flow

clear_daily_flow

deer_daily_flow

rbdd_daily_temp

feather_daily_temp

tuol_daily_temp

amer_daily_temp

moke_daily_temp

stan_daily_temp

clear_daily_temp

deer_daily_temp

Format

An object of class tbl_df (inherits from tbl, data.frame) with 216 rows and 3 columns.

Details

Available screw trap data start and end years:

Location 	 start-end

RBDD 	 1999 - 2016

Feather River 	 1998 - 2016

Tuolumne River 	 2007 - 2017

American River 	 2013 - 2016

Battle Creek 	 1998 - 2016

Clear Creek 	 1998 - 2016

Deer Creek 	 1992 - 2010

Mokelumne River 	 1999 - 2015

Flow and Temperature Gage IDs

Flow Gages:

RBDD (Upper Sacramento River) - USGS SACRAMENTO R AB BEND BRIDGE NR RED BLUFF CA (11377100)

Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)

Tuolumne River - USGS TUOLUMNE R A MODESTO CA (11290000)

American River - USGS AMERICAN R A FAIR OAKS CA (11446500)

Battle Creek - USGS BATTLE C BL COLEMAN FISH HATCHERY NR COTTONWOOD CA (11376550)

Clear Creek - CDEC CLEAR CREEK NEAR IGO (IGO)

Deer Creek - USGS DEER C NR VINA CA (11383500)

Mokelumne River - USGS MOKELUMNE R BL CAMANCHE DAM CA (11323500)

Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)

Temperature Gages:

RBDD (Upper Sacramento River) - CDEC SACRAMENTO R AT RED BLUFF DIVERSION DAM (RDB)

Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)

Tuolumne River - USGS TUOLUMNE R A MODESTO CA (11290000)

American River - USGS AMERICAN R A FAIR OAKS CA (11446500)

Battle Creek - USGS BATTLE C BL COLEMAN FISH HATCHERY NR COTTONWOOD CA (11376550)

Clear Creek - CDEC CLEAR CREEK NEAR IGO (IGO)

Deer Creek - USGS DEER C NR VINA CA (11383500)

Mokelumne River - EDMUD provided data near Victor CA

Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)

NOTE: Deer Creek temperature data begins in 1998

Daily Data

Daily data for flow and temperature are also available: [watershed]_daily_flow or [watershed]_daily_temp