Synthetic Years Model inputs

All the model inputs for sythetic years to cover 1998-2017

fall_inputs

winter_inputs

spring_inputs

steelhead_inputs

Format

list

Details

Available screw trap data start and end years:

Location	start-end

Feather River	1998 - 2016

American River	2013 - 2016

Clear Creek	1998 - 2016

Mokelumne River	1999 - 2015

Stanislaus River	1998 - 2016

NOTE:

Habitat values were scaled using the proportion of spawning and rearing extent above the trap.

Apply a suitability factor to floodplain areas (suggested value = .27).