Proportion and Total Diverted

Monthly propotion and total diverted values for the duration of screw trap observations.

feat_prop_div

amer_prop_div

moke_prop_div

stan_prop_div

feat_tot_div

amer_tot_div

moke_tot_div

stan_tot_div

Format

An object of class tbl_df (inherits from tbl, data.frame) with 240 rows and 3 columns.

Details

Available screw trap data start and end years:

Location	start-end

Feather River	1998 - 2016

American River	2013 - 2016

Clear Creek	1998 - 2016

Mokelumne River	1999 - 2015

Stanislaus River	1998 - 2016

Proportion Diverted

Values pre September 2003 are CALSIM model output and values after are the median CALSIM value grouped by water year type and month.

For more details on CALSIM see:

use this link within R proportion_diverted

use this link if in a web browser

NOTE: No major diversions to track for Clear Creek.

Mokelumne data is measured provided by EDMUD

Total Diverted

Values were estimated by multiplying the measured flow data and the modeled proportion diverted values.

NOTE: Mokelumne data is measured provided by EDMUD